+++ Assets/Art/Materials/Beans_Tube_Deform1.mat
@@ -1,0 +1,90 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &-2585888148910458892
+MonoBehaviour:
+  m_ObjectHideFlags: 11
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d0353a89b1f911e48b9e16bdc9f2e058, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  version: 5
+--- !u!21 &2100000
+Material:
+  serializedVersion: 8
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Beans_Tube_Deform1
+  m_Shader: {fileID: -6465566751694194690, guid: bb997e27d5208ef49ad75130dcd9ebb7, type: 3}
+  m_ValidKeywords: []
+  m_InvalidKeywords: []
+  m_LightmapFlags: 2
+  m_EnableInstancingVariants: 0
+  m_DoubleSidedGI: 0
+  m_CustomRenderQueue: -1
+  stringTagMap:
+    RenderType: Opaque
+  disabledShaderPasses: []
+  m_SavedProperties:
+    serializedVersion: 3
+    m_TexEnvs:
+    - _SampleTexture2DLOD_445e99616dfd4b57bc2032992bb62cae_Texture_1:
+        m_Texture: {fileID: 2800000, guid: 8c0ce78c5eb9e1d4fabebb2fb5f38e70, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _Texture:
+        m_Texture: {fileID: 2800000, guid: d926d3d80e7e12048ad20e9c2a61451d, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _Texture2DAsset_f663c0e9b84841e38bc7af88d48a4be1_Out_0:
+        m_Texture: {fileID: 2800000, guid: 8c0ce78c5eb9e1d4fabebb2fb5f38e70, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _buldge_shape:
+        m_Texture: {fileID: 2800000, guid: 7d8aa960ca19dcf489a521b7cb71a1d4, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - unity_Lightmaps:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - unity_LightmapsInd:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - unity_ShadowMasks:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    m_Ints: []
+    m_Floats:
+    - _AlphaClip: 0
+    - _Blend: 0
+    - _Buldge_strength: 0.007
+    - _CastShadows: 1
+    - _Cull: 2
+    - _Deformation_limit: 1.5
+    - _DstBlend: 0
+    - _Edge_Buffer_strength: 20
+    - _QueueControl: 0
+    - _QueueOffset: 0
+    - _ReceiveShadows: 1
+    - _SrcBlend: 1
+    - _Surface: 0
+    - _WorkflowMode: 1
+    - _ZTest: 4
+    - _ZWrite: 1
+    - _ZWriteControl: 0
+    - _tileing: 2
+    m_Colors:
+    - _BaseColor: {r: 0.34509805, g: 0, b: 0.4392157, a: 1}
+    - _Base_Color: {r: 0.7735849, g: 0, b: 0, a: 0}
+    - _Buldge_Color: {r: 8, g: 2.1960785, b: 0, a: 0}
+    - _Scroll_speed_Direction: {r: -0.7, g: 0.5, b: 0, a: 0}
+  m_BuildTextureStacks: []
