+++ Assets/Art/Materials/Beans_Tube_Deform1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4214151836e3cba4188a490d956210c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
