+++ Assets/Art/Materials/Beans_Tube_Deform2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 16237bf9d1adac641851a572e71cadfc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
