+++ Assets/Art/Materials/Bio_Tube_Deform1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4abc3eb0de145634891084676b015530
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
