+++ Assets/Art/Materials/Bio_Tube_Deform2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0a69b746a0663846a27422433861417
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
