+++ Assets/Art/Materials/Liquid_Tube_Deform1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d2d0e41046659d41a67ca6d254d971e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
