+++ Assets/Art/Materials/Liquid_Tube_Deform2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 943621a1b98a01f4e8c044048a89730b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
