+++ Assets/Art/Materials/Sweet_Tube_Deform2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69874278b5cfe5042abfad1211f37e0b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
