+++ Assets/Art/Materials/sweet_Tube_Deform1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 417b00313e3eea04c8a7c31c83f277d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
