+++ Assets/Art/VFX/Samoris_work/Full screen VFX/WIFI down effect.mat
@@ -61,7 +61,7 @@ Material:
     - _DstBlend: 10
     - _Effect_Speed: 1
     - _Effect_power: 8
-    - _Effect_scale: 500
+    - _Effect_scale: 623.7
     - _Full_screen_effect_Intensity: 0.1
     - _QueueControl: 0
     - _QueueOffset: 0
@@ -74,5 +74,5 @@ Material:
     - _ZWriteControl: 0
     m_Colors:
     - _Color: {r: 103.449196, g: 103.449196, b: 103.449196, a: 1}
-    - _Effect_Speed: {r: 15, g: 0.1, b: 0, a: 0}
+    - _Effect_Speed: {r: 0.5, g: 10, b: 0, a: 0}
   m_BuildTextureStacks: []
