--- Assets/Art/VFX/Samoris_work/Pipe deformation/Shader Graphs_test graph 1.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 16237bf9d1adac641851a572e71cadfc
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
