--- Assets/Art/VFX/Samoris_work/Pipe deformation/Shader Graphs_test graph.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4214151836e3cba4188a490d956210c5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
