+++ Assets/Art/VFX/Samoris_work/Pipe deformation/test graph.shadergraph
@@ -2130,9 +2130,9 @@
         "m_Expanded": true,
         "m_Position": {
             "serializedVersion": "2",
-            "x": 3458.0,
-            "y": -382.0,
-            "width": 105.0,
+            "x": 3463.000244140625,
+            "y": -543.0,
+            "width": 145.0,
             "height": 34.0
         }
     },
