+++ Assets/Art/VFX/Samoris_work/Pipe deformation/tube material temp.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6aee3429db7ee564fa8c55dc3bf40492
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
