+++ Assets/Art/VFX/Samoris_work/ping effect/Ping Icons/BuldgeControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce509968d431bf24dace7ac4f6a96842
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
