+++ Assets/Art/station animations/Samori's animations/Animations/Bean station animations/Beanstation7.playable
@@ -81,17 +81,17 @@ AnimationClip:
       m_Curve:
       - serializedVersion: 3
         time: 0
-        value: {x: 4.361779, y: -0.02248487, z: -0.001305775}
+        value: {x: 4.361779, y: -0.02248487, z: 89.7067}
         inSlope: {x: 0, y: 0, z: 0}
-        outSlope: {x: 90.08585, y: 0, z: 0}
+        outSlope: {x: 90.085884, y: 0, z: 0}
         tangentMode: 0
         weightedMode: 0
         inWeight: {x: 0.33333334, y: 0.33333334, z: 0.33333334}
         outWeight: {x: 0.33333334, y: 0.33333334, z: 0.33333334}
       - serializedVersion: 3
         time: 8
-        value: {x: 725.0486, y: -0.02247618, z: -0.001312251}
-        inSlope: {x: 90.08585, y: 0, z: 0}
+        value: {x: 725.0488, y: -0.0219879, z: 89.70551}
+        inSlope: {x: 90.085884, y: 0, z: 0}
         outSlope: {x: 0, y: 0, z: 0}
         tangentMode: 0
         weightedMode: 0
@@ -171,15 +171,15 @@ AnimationClip:
         time: 0
         value: 4.361779
         inSlope: 0
-        outSlope: 90.08585
+        outSlope: 90.085884
         tangentMode: 69
         weightedMode: 0
         inWeight: 0.33333334
         outWeight: 0.33333334
       - serializedVersion: 3
         time: 8
-        value: 725.0486
-        inSlope: 90.08585
+        value: 725.0488
+        inSlope: 90.085884
         outSlope: 0
         tangentMode: 69
         weightedMode: 0
@@ -206,7 +206,7 @@ AnimationClip:
         outWeight: 0.33333334
       - serializedVersion: 3
         time: 8
-        value: -0.02247618
+        value: -0.0219879
         inSlope: 0
         outSlope: 0
         tangentMode: 136
@@ -225,7 +225,7 @@ AnimationClip:
       m_Curve:
       - serializedVersion: 3
         time: 0
-        value: -0.001305775
+        value: 89.7067
         inSlope: 0
         outSlope: 0
         tangentMode: 136
@@ -234,7 +234,7 @@ AnimationClip:
         outWeight: 0.33333334
       - serializedVersion: 3
         time: 8
-        value: -0.001312251
+        value: 89.70551
         inSlope: 0
         outSlope: 0
         tangentMode: 136
