+++ Assets/Development/Prefabs/Stations/Ingredient/LiquidStation.prefab
@@ -5284,9 +5284,9 @@ PlayableDirector:
   m_GameObject: {fileID: 4175465839301593472}
   m_Enabled: 1
   serializedVersion: 3
-  m_PlayableAsset: {fileID: 11400000, guid: 9c8c28c39cc92f34a999af7e48b885ea, type: 2}
+  m_PlayableAsset: {fileID: 11400000, guid: 73284ad8e274b824ba576087bb040e8d, type: 2}
   m_InitialState: 1
-  m_WrapMode: 2
+  m_WrapMode: 1
   m_DirectorUpdateMode: 1
   m_InitialTime: 0
   m_SceneBindings:
@@ -8377,7 +8377,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7203993366632014359, guid: 81393d6a781120545850ed7b0e4c269a, type: 3}
       propertyPath: m_Enabled
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7203993366632021480, guid: 81393d6a781120545850ed7b0e4c269a, type: 3}
       propertyPath: m_Name
@@ -8911,7 +8911,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7203993366632014359, guid: 81393d6a781120545850ed7b0e4c269a, type: 3}
       propertyPath: m_Enabled
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7203993366632021480, guid: 81393d6a781120545850ed7b0e4c269a, type: 3}
       propertyPath: m_Name
@@ -8963,7 +8963,7 @@ PrefabInstance:
     - target: {fileID: -8698591140399667973, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
       propertyPath: m_Materials.Array.data[0]
       value: 
-      objectReference: {fileID: 2100000, guid: 52606f77732d7054b8d5bee83141a310, type: 2}
+      objectReference: {fileID: 2100000, guid: 943621a1b98a01f4e8c044048a89730b, type: 2}
     - target: {fileID: -8679921383154817045, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
       propertyPath: m_RootOrder
       value: 1
@@ -9091,7 +9091,7 @@ PrefabInstance:
     - target: {fileID: 6368978175081929303, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
       propertyPath: m_Materials.Array.data[0]
       value: 
-      objectReference: {fileID: 2100000, guid: b2f82e6a8e26a6a40ac8138060235950, type: 2}
+      objectReference: {fileID: 2100000, guid: 6d2d0e41046659d41a67ca6d254d971e, type: 2}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
 --- !u!4 &1110883750501645389 stripped
@@ -9136,13 +9136,29 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2410085642011316664}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 31699402b3ee0bf4b82f414819b18df3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   emissiveMaterial: {fileID: 2100000, guid: b2f82e6a8e26a6a40ac8138060235950, type: 2}
   emissionIntensity: 1
+--- !u!114 &277367686
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2410085642011316664}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ce509968d431bf24dace7ac4f6a96842, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  buldgeMaterial: {fileID: 2100000, guid: 6d2d0e41046659d41a67ca6d254d971e, type: 2}
+  buldgeColor: {r: 0, g: 1, b: 0.9319372, a: 1}
+  emissionIntensity: 3
+  buldgeStrength: 0.007
 --- !u!1 &4227509758758026756 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 3624058657922016674, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
@@ -9255,13 +9271,29 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8719248784560523440}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 31699402b3ee0bf4b82f414819b18df3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   emissiveMaterial: {fileID: 2100000, guid: 52606f77732d7054b8d5bee83141a310, type: 2}
   emissionIntensity: 1
+--- !u!114 &1759459444
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8719248784560523440}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ce509968d431bf24dace7ac4f6a96842, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  buldgeMaterial: {fileID: 2100000, guid: 943621a1b98a01f4e8c044048a89730b, type: 2}
+  buldgeColor: {r: 0, g: 1, b: 0.9319372, a: 1}
+  emissionIntensity: 3
+  buldgeStrength: 0.007
 --- !u!1 &9079518655325841288 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 8566112037541821486, guid: 01f21dde5f02ae0449d7dc3298fc6d96, type: 3}
